refactor:move controller class
